Move ArchiveTask to bots
